% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_method.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Extract regression coefficients}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
A named vector of regression coefficients.
}
\description{
Returns the estimated coefficients from a linreg object.
}
