% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A Reference Class to represent a linear regression package in R}
\arguments{
\item{m_data}{}
}
\value{
Nothing
}
\description{
Listed functions are as methods in the object included:

- print() prints out the coefficients and coefficient names, similar as done by the lm class

- plot() plots the two plots using ggplot2

1. plot "Residuals vs Fitted"
2. plot "Scale−Location"

- resid() returns the vector of residuals e

- pred() returns the predicted values y

- coef() returns the coefficients as a named vector.

- summary() returns a similar printout as printed for lm objects,

- but only presents the coefficients with their standard error,

- t-value and p-value as well as the estimate of σˆ and the degrees of freedom in the model.
}
\section{Fields}{

\describe{
\item{\code{X}}{the values of X as matrix}

\item{\code{y}}{the values of y as matrix}

\item{\code{regressions_coef}}{the regression coefficients, saved as matrix}

\item{\code{fitted_values}}{the fitted values, saved as matrix}

\item{\code{resi}}{the residuals, saved as matrix}

\item{\code{n}}{the number of observations, saged as integer}

\item{\code{p}}{the number of parameters in the model, saved as integer}

\item{\code{dof}}{the degree of freedom, saved as integer}

\item{\code{regressions_var}}{the variance of the regression coefficent, saved as matrix}

\item{\code{t_value}}{is the t-value, saved as matrix}
}}


