% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Fit a linear regression model}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{The formula that describes the regression model.}

\item{data}{The data frame used to fit the model.}
}
\value{
A "linreg" object with the estimated coefficients,
fitted values, residuals and other model results.
}
\description{
Creates a multiple linear regression model based on a given formula
and data set.
}
\examples{
model <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, iris)

}
