% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_qr.R
\name{linreg_qr}
\alias{linreg_qr}
\title{Linear regression using QR decomposition}
\usage{
linreg_qr(formula, data)
}
\arguments{
\item{formula}{A formula defining the response and predictor variables.}

\item{data}{A data frame containing the variables used in the model.}
}
\value{
An object of class "linreg_qr" containing the regression results.
}
\description{
Fits a multiple linear regression model using QR decomposition
to calculate the regression coefficients and their variances.
}
\examples{
fit <- linreg_qr(
  Petal.Length ~ Sepal.Width + Sepal.Length,
  data = iris
)

}
