% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{my_print}
\alias{my_print}
\title{my_print function
Our own print function, because we can't call the default function inside RC object}
\usage{
my_print(x)
}
\arguments{
\item{x}{is the input.}
}
\description{
my_print function
Our own print function, because we can't call the default function inside RC object
}
\examples{
print("This is an example!")
}
