% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_method.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot a linreg object}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
Two diagnostic plots.
}
\description{
Creates diagnostic plots for a fitted linear regression model.
}
