% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_method.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Extract predicted values from a linreg object}
\usage{
\method{pred}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
A numeric vector of predicted values.
}
\description{
Returns the fitted values from a linreg object.
}
