% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_method.R
\name{pred}
\alias{pred}
\title{Extract predicted values}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Not used.}
}
\value{
Predicted values.
}
\description{
Generic function for extracting predicted values.
}
