% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_method.R
\name{print.linreg}
\alias{print.linreg}
\title{Print a linreg object}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
The linreg object invisibly.
}
\description{
Prints the model call and estimated coefficients.
}
