% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_method.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summarize a linreg object}
\usage{
\method{summary}{linreg}(x)
}
\arguments{
\item{x}{An object of class "linreg".}
}
\value{
A summary of the fitted regression model.
}
\description{
Prints the coefficient estimates, standard errors, t-values,
p-values and residual standard error.
}
