% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{write_star}
\alias{write_star}
\title{write_star function
generate the start base on p_value, follow the rule of lm function}
\usage{
write_star(p_value)
}
\arguments{
\item{p_value}{the input}
}
\description{
write_star function
generate the start base on p_value, follow the rule of lm function
}
