% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_method.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Extract regression coefficients}
\usage{
\method{coef}{linreg}(x)
}
\arguments{
\item{x}{An object of class "linreg".}
}
\value{
A named vector of regression coefficients.
}
\description{
Returns the estimated coefficients from a linreg object.
}
