% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_method.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{Extract residuals}
\usage{
\method{residuals}{linreg}(x)
}
\arguments{
\item{x}{An object of class "linreg".}
}
\value{
A numeric vector of residuals.
}
\description{
Returns the residuals from a linreg object.
}
