% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_method.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summarize a linreg object}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object of class "linreg".}

\item{...}{Not used.}
}
\value{
A summary of the fitted regression model.
}
\description{
Prints the coefficient estimates, standard errors, t-values,
p-values and residual standard error.
}
